# openapi.yaml
openapi: 3.1.0
info:
  title: Marathon LFG API
  version: 0.2.0
  description: |
    Backend contract for the Marathon Looking For Group MVP.

    Browser/PWA notes:
    - Web auth starts with `POST /auth/bungie/start` using `redirectMode: web`.
    - Web callback sets the `mlfg_at`, `mlfg_rt`, and `mlfg_csrf` cookies.
    - Cookie-authenticated mutation requests must send `X-CSRF-Token` matching `mlfg_csrf`.
    - The browser shell is served from `/app`, but the shell routes themselves are not modeled as API operations here.
servers:
  - url: http://127.0.0.1:3000
    description: Local development
...
paths:
...
  /parties:
    get:
      tags: [Parties]
      summary: List visible parties
      security:
        - MlfgAccessCookie: []
        - BearerAuth: []
        - {}
      responses:
        '200':
          description: Public parties plus viewer-visible private/cancelled parties.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PartySummary'
                required: [items]
    post:
      tags: [Parties]
      summary: Create a party
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      security:
        - BearerAuth: []
        - MlfgAccessCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartyRequest'
      responses:
        '201':
          description: Party created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyCreateResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /parties/{partyId}:
    get:
      tags: [Parties]
      summary: Get party detail
      security:
        - MlfgAccessCookie: []
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/PartyId'
      responses:
        '200':
          description: Party detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyDetail'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    patch:
      tags: [Parties]
      summary: Edit a party
      description: Host-only party metadata edit.
      parameters:
        - $ref: '#/components/parameters/PartyId'
        - $ref: '#/components/parameters/CsrfHeader'
      security:
        - BearerAuth: []
        - MlfgAccessCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePartyRequest'
      responses:
        '200':
          description: Party updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyDetail'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /parties/{partyId}/join:
    post:
      tags: [Parties]
      summary: Request to join a party
      parameters:
        - $ref: '#/components/parameters/PartyId'
        - $ref: '#/components/parameters/CsrfHeader'
      security:
        - BearerAuth: []
        - MlfgAccessCookie: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinPartyRequest'
      responses:
        '200':
          description: Join request accepted or queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyJoinResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /parties/{partyId}/leave:
    post:
      tags: [Parties]
      summary: Leave a party
      parameters:
        - $ref: '#/components/parameters/PartyId'
        - $ref: '#/components/parameters/CsrfHeader'
      security:
        - BearerAuth: []
        - MlfgAccessCookie: []
      responses:
        '200':
          description: Membership moved to `left`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyMutationResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /parties/{partyId}/cancel:
    post:
      tags: [Parties]
      summary: Cancel a hosted party
      parameters:
        - $ref: '#/components/parameters/PartyId'
        - $ref: '#/components/parameters/CsrfHeader'
      security:
        - BearerAuth: []
        - MlfgAccessCookie: []
      responses:
        '200':
          description: Party cancelled.
          content:
            application/json:
              schema:
                type: object
                properties:
                  partyId:
                    type: string
                    format: uuid
                  status:
                    $ref: '#/components/schemas/PartyStatus'
                required: [partyId, status]
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /parties/{partyId}/members/{memberId}/accept:
    post:
      tags: [Parties]
      summary: Accept a pending member
      parameters:
        - $ref: '#/components/parameters/PartyId'
        - $ref: '#/components/parameters/MemberId'
        - $ref: '#/components/parameters/CsrfHeader'
      security:
        - BearerAuth: []
        - MlfgAccessCookie: []
      responses:
        '200':
          description: Member accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyMutationResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /parties/{partyId}/members/{memberId}/decline:
    post:
      tags: [Parties]
      summary: Decline a pending member
      parameters:
        - $ref: '#/components/parameters/PartyId'
        - $ref: '#/components/parameters/MemberId'
        - $ref: '#/components/parameters/CsrfHeader'
      security:
        - BearerAuth: []
        - MlfgAccessCookie: []
      responses:
        '200':
          description: Member declined.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyMutationResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /parties/{partyId}/members/{memberId}/kick:
    post:
      tags: [Parties]
      summary: Kick a pending or accepted member
      parameters:
        - $ref: '#/components/parameters/PartyId'
        - $ref: '#/components/parameters/MemberId'
        - $ref: '#/components/parameters/CsrfHeader'
      security:
        - BearerAuth: []
        - MlfgAccessCookie: []
      responses:
        '200':
          description: Member kicked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyMutationResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Native/mobile style bearer access token.
    MlfgAccessCookie:
      type: apiKey
      in: cookie
      name: mlfg_at
      description: Browser/PWA access-token cookie. Non-GET requests also need `X-CSRF-Token`.
  parameters:
    PartyId:
      name: partyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: string
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required when authenticating mutation requests with browser cookies.
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    CreatePartyRequest:
      type: object
      additionalProperties: false
      required: [title, activityKey, maxSize]
      properties:
        title:
          type: string
        activityKey:
          type: string
          description: Current PWA client uses this for map selection so party lists can filter by map.
          example: perimeter
        playlistKey:
          type: string
          description: Current PWA client uses this for mode selection, currently "duos" or "trios".
          example: trios
        platformKey:
          type: string
        regionKey:
          type: string
        languageKey:
          type: string
        voiceRequired:
          type: boolean
        ranked:
          type: boolean
        scheduledFor:
          type: string
        maxSize:
          type: integer
          minimum: 1
          description: Current PWA client derives this from playlistKey; duos uses 2 and trios uses 3.
        approvalMode:
          type: string
        visibility:
          type: string
        requiresMarathonVerified:
          type: boolean
        requirementText:
          type: string
        description:
          type: string
        externalJoinUrl:
          type: string
        tags:
          type: array
          description: Current PWA client sends non-map preferences, such as shell, as tags.
          example:
            - tagKey: shell
              tagValue: recon
          items:
            $ref: '#/components/schemas/PartyTag'
    UpdatePartyRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
        activityKey:
          type: string
          description: Current PWA client uses this for map selection so party lists can filter by map.
          example: perimeter
        playlistKey:
          type: [string, 'null']
          description: Current PWA client uses this for mode selection, currently "duos" or "trios".
          example: trios
        platformKey:
          type: string
        regionKey:
          type: [string, 'null']
        languageKey:
          type: [string, 'null']
        voiceRequired:
          type: boolean
        ranked:
          type: [boolean, 'null']
        scheduledFor:
          type: [string, 'null']
        maxSize:
          type: integer
          minimum: 1
          description: Current PWA client derives this from playlistKey; duos uses 2 and trios uses 3.
        approvalMode:
          type: string
        visibility:
          type: string
        requiresMarathonVerified:
          type: boolean
        requirementText:
          type: [string, 'null']
        description:
          type: [string, 'null']
        externalJoinUrl:
          type: [string, 'null']
        tags:
          type: array
          description: Replaces the full tag set for the party when provided.
          items:
            $ref: '#/components/schemas/PartyTag'
    PartyCreateResponse:
      type: object
      properties:
        partyId:
          type: string
          format: uuid
        status:
          type: string
        filledSlots:
          type: integer
        openSlots:
          type: integer
      required: [partyId, status, filledSlots, openSlots]
    JoinPartyRequest:
      type: object
      additionalProperties: false
      properties:
        noteToHost:
          type: string
    PartyJoinResponse:
      type: object
      properties:
        partyId:
          type: string
          format: uuid
        myStatus:
          $ref: '#/components/schemas/PartyMemberStatus'
        filledSlots:
          type: integer
        openSlots:
          type: integer
      required: [partyId, myStatus, filledSlots, openSlots]
    PartyMutationResponse:
      type: object
      properties:
        partyId:
          type: string
          format: uuid
        memberId:
          type: string
        memberStatus:
          $ref: '#/components/schemas/PartyMemberStatus'
        filledSlots:
          type: integer
        openSlots:
          type: integer
        partyStatus:
          type: string
      required: [partyId, memberId, memberStatus, filledSlots, openSlots, partyStatus]
    PartySummary:
      type: object
      properties:
        partyId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PartyStatus'
        title:
          type: string
        activityKey:
          type: string
          example: perimeter
        playlistKey:
          type: [string, 'null']
          example: trios
        platformKey:
          type: string
        regionKey:
          type: [string, 'null']
        languageKey:
          type: [string, 'null']
        voiceRequired:
          type: boolean
        ranked:
          type: [boolean, 'null']
        scheduledFor:
          type: [string, 'null']
        maxSize:
          type: integer
          description: Current PWA client derives this from playlistKey; duos uses 2 and trios uses 3.
        approvalMode:
          type: string
        visibility:
          type: string
        requiresMarathonVerified:
          type: boolean
        requirementText:
          type: [string, 'null']
        description:
          type: [string, 'null']
        externalJoinUrl:
          type: [string, 'null']
        filledSlots:
          type: integer
        openSlots:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        host:
          $ref: '#/components/schemas/PartyPerson'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/PartyTag'
        myMembership:
          oneOf:
            - $ref: '#/components/schemas/PartyMembershipSummary'
            - type: 'null'
        members:
          type: array
          items:
            $ref: '#/components/schemas/PartyMemberView'
      required:
        - partyId
        - status
        - title
        - activityKey
        - playlistKey
        - platformKey
        - regionKey
        - languageKey
        - voiceRequired
        - ranked
        - scheduledFor
        - maxSize
        - approvalMode
        - visibility
        - requiresMarathonVerified
        - requirementText
        - description
        - externalJoinUrl
        - filledSlots
        - openSlots
        - createdAt
        - updatedAt
        - host
        - tags
        - myMembership
        - members
    PartyDetail:
      allOf:
        - $ref: '#/components/schemas/PartySummary'
    PartyTag:
      type: object
      description: Free-form party metadata. Current PWA conventions include tagKey=shell. Map is stored in activityKey for filtering.
      properties:
        tagKey:
          type: string
          example: shell
        tagValue:
          type: [string, 'null']
          example: recon
      required: [tagKey, tagValue]
    PartyPerson:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        bungieDisplayName:
          type: [string, 'null']
        globalDisplayName:
          type: [string, 'null']
        globalDisplayNameCode:
          type: [integer, 'null']
      required: [userId, bungieDisplayName, globalDisplayName, globalDisplayNameCode]
    PartyMembershipSummary:
      type: object
      properties:
        memberId:
          type: string
        status:
          $ref: '#/components/schemas/PartyMemberStatus'
        noteToHost:
          type: [string, 'null']
        joinedAt:
          type: [string, 'null']
        respondedAt:
          type: [string, 'null']
        createdAt:
          type: string
      required: [memberId, status, noteToHost, joinedAt, respondedAt, createdAt]
    PartyMemberView:
      type: object
      properties:
        memberId:
          type: string
        userId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PartyMemberStatus'
        noteToHost:
          type: [string, 'null']
        joinedAt:
          type: [string, 'null']
        respondedAt:
          type: [string, 'null']
        createdAt:
          type: string
        bungieDisplayName:
          type: [string, 'null']
        globalDisplayName:
          type: [string, 'null']
        globalDisplayNameCode:
          type: [integer, 'null']
      required:
        - memberId
        - userId
        - status
        - noteToHost
        - joinedAt
        - respondedAt
        - createdAt
        - bungieDisplayName
        - globalDisplayName
        - globalDisplayNameCode
    PartyStatus:
      type: string
      enum: [open, full, in_progress, closed, cancelled]
    PartyMemberStatus:
      type: string
      enum: [accepted, pending, declined, left, kicked]